Extract lightmap/shading/geometry helpers into src/app/render.js
Moves the pure rendering math and helpers (buildHillshadeQ, buildLightmap,
sampleLightAt, ensureLightmapBuffers, maybeBuildLightmap, emittersSignature,
shadeFillColor + applyNightColorShift, applySpriteShade, *Lit variants,
entityDrawRect, visibleTileBounds, lightmapCacheState, resetLightmapCache,
currentAmbient) into a dedicated module exposed via createRenderSystem.

The state-mutating frame pipeline (render(), drawXxx entity helpers,
zone/water overlays, nocturnal entities, drawShadow, drawStaticAlbedo)
remains in app.js for now — those have heavy coupling to in-flight game
state and would need a much larger surface refactor to relocate cleanly.
